+++ app/src/main/java/com/sunloto/drawing/lotterydrawresult/bean/ToWinToNil.java
@@ -15,7 +15,7 @@ public class ToWinToNil implements Parcelable{
     @SerializedName(FIELD_TEAM)
     private String mTeam;
     @SerializedName(FIELD_ODDS)
-    private int mOdd;
+    private double mOdd;
     @SerializedName(FIELD_TITLE)
     private String mTitle;
 
@@ -32,11 +32,11 @@ public class ToWinToNil implements Parcelable{
         return mTeam;
     }
 
-    public void setOdd(int odd) {
+    public void setOdd(double odd) {
         mOdd = odd;
     }
 
-    public int getOdd() {
+    public double getOdd() {
         return mOdd;
     }
 
@@ -50,7 +50,7 @@ public class ToWinToNil implements Parcelable{
 
     public ToWinToNil(Parcel in) {
         mTeam = in.readString();
-        mOdd = in.readInt();
+        mOdd = in.readDouble();
         mTitle = in.readString();
     }
 
@@ -72,7 +72,7 @@ public class ToWinToNil implements Parcelable{
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeString(mTeam);
-        dest.writeInt(mOdd);
+        dest.writeDouble(mOdd);
         dest.writeString(mTitle);
     }
